More Clang Changes
Clang complains about large ints auto-casting to float.
Moreover, clang changes the flag to supress that warning
from clang 9 -> clang 10.

So just put in explicit casts to float.

Addresses #2386

+++ src/common/dsp/DspUtilities.cpp
@@ -4,7 +4,7 @@ float correlated_noise(float lastval, float correlation)
 {
    float wf = correlation * 0.9;
    float wfabs = fabs(wf);
-   float rand11 = (((float)rand() / RAND_MAX) * 2.f - 1.f);
+   float rand11 = (((float)rand() / (float)RAND_MAX) * 2.f - 1.f);
    float randt = rand11 * (1 - wfabs) - wf * lastval;
    return randt;
 }
@@ -14,7 +14,7 @@ float correlated_noise_mk2(float& lastval, float correlation)
    float wf = correlation * 0.9;
    float wfabs = fabs(wf);
    float m = 1.f / sqrt(1.f - wfabs);
-   float rand11 = (((float)rand() / RAND_MAX) * 2.f - 1.f);
+   float rand11 = (((float)rand() / (float)RAND_MAX) * 2.f - 1.f);
    lastval = rand11 * (1 - wfabs) - wf * lastval;
    return lastval * m;
 }
@@ -26,7 +26,7 @@ float drift_noise(float& lastval)
    //__m128 mvec = _mm_rsqrt_ss(_mm_load_ss(&filter));
    //_mm_store_ss(&m,mvec);
 
-   float rand11 = (((float)rand() / RAND_MAX) * 2.f - 1.f);
+   float rand11 = (((float)rand() / (float)RAND_MAX) * 2.f - 1.f);
    lastval = lastval * (1.f - filter) + rand11 * filter;
    return lastval * m;
 }
@@ -35,7 +35,7 @@ float correlated_noise_o2(float lastval, float& lastval2, float correlation)
 {
    float wf = correlation * 0.9;
    float wfabs = fabs(wf);
-   float rand11 = (((float)rand() / RAND_MAX) * 2.f - 1.f);
+   float rand11 = (((float)rand() / (float)RAND_MAX) * 2.f - 1.f);
    float randt = rand11 * (1 - wfabs) - wf * lastval2;
    lastval2 = randt;
    randt = lastval2 * (1 - wfabs) - wf * lastval;
@@ -61,7 +61,7 @@ float correlated_noise_o2mk2(float& lastval, float& lastval2, float correlation)
    _mm_store_ss(&m, m1);
    // if (wf>0.f) m *= 1 + wf*8;
 #endif
-   float rand11 = (((float)rand() / RAND_MAX) * 2.f - 1.f);
+   float rand11 = (((float)rand() / (float)RAND_MAX) * 2.f - 1.f);
    lastval2 = rand11 * (1 - wfabs) - wf * lastval2;
    lastval = lastval2 * (1 - wfabs) - wf * lastval;
    return lastval * m;
